+++ src/Fable.Cli/RELEASE_NOTES.md
@@ -1,3 +1,9 @@
+### 3.0.0-nagareyama-alpha-003
+
+* Fix: Add names with $reflection suffix to root scope
+* Fix deriving from imported JS classes
+* Keep support for ParamList attribute until removed from Fable.React
+
 ### 3.0.0-nagareyama-alpha-002
 
 * Ignore warnings from packages (as before)